pyproject.toml
[project]
name = "score_core"
version = "0.1.0"
description = "SCORE: Story Coherence and Retrieval Enhancement core library"
authors = [{ name = "Sari", email = "6322102ys@gmail.com" }]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.7",
    "faiss-cpu>=1.8",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "openai>=1.14"        # Consistency 判定で使用
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
